操作

+++ GLFWorkspace/GLFiTunesShare/GLFiTunesShare/controller/Root/SubViewController2.m
@@ -93,7 +93,7 @@
 }
 
 // 完成放大缩小时调用
-- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(float)scale; {
+- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale; {
     if (scale>=minScale && scale<=maxScale) {
         currentScale = scale;
     }
